requirements.txt
absl-py==2.0.0
accelerate==0.21.0
addict==2.4.0
aenum==3.1.15
aiofiles==23.1.0
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiosignal==1.3.1
altair==5.0.1
altgraph==0.17.2
aniso8601==9.0.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argilla==1.12.0
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
asyncssh==2.13.2
attrs==23.1.0
audiolm-pytorch==1.0.4
audioread==3.0.1
Automat==20.2.0
autopep8==1.6.0
av==11.0.0
awscli==1.29.41
azure-cognitiveservices-speech==1.22.0
backoff==2.2.1
basicsr==1.4.2
bcrypt==4.0.1
beartype==0.14.0
beautifulsoup4==4.11.1
black==23.3.0
blendmodes==2022
blessed==1.20.0
blinker==1.5
boltons==23.1.1
boto3==1.28.41
botocore==1.31.41
bs4==0.0.1
cachetools==5.3.0
certifi==2021.10.8
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==2.0.12
chromadb==0.3.22
ci-info==0.3.0
clean-fid==0.1.35
click==8.0.3
clickhouse-connect==0.5.23
clip @ https://github.com/openai/CLIP/archive/d50d76daa670286dd6cacf3bcd80b5e4823fc8e1.zip#sha256=b5842c25da441d6c581b53a5c60e0c2127ebafe0f746f8e15561a006c6c3be6a
colorama==0.4.4
coloredlogs==15.0.1
colorful==0.5.5
colorlog==6.8.0
commonmark==0.9.1
configobj==5.0.8
configparser==6.0.0
constantly==15.1.0
contourpy==1.1.0
coverage==7.2.3
cryptography==38.0.1
cssselect==1.1.0
cssselect2==0.7.0
cycler==0.11.0
dataclasses-json==0.5.7
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
distlib==0.3.4
dnspython==2.3.0
docker==6.0.1
docutils==0.16
docx2txt==0.8
duckdb==0.7.1
duckduckgo-search==2.8.6
einops==0.4.1
et-xmlfile==1.1.0
etelemetry==0.3.0
exceptiongroup==1.0.4
face-alignment==1.3.5
facexlib==0.3.0
faiss-cpu==1.7.4
fastapi==0.94.0
feedparser==6.0.10
ffmpy==0.3.0
filelock==3.6.0
filetype==1.2.0
filterpy==1.4.5
fitz==0.0.1.dev2
flake8==6.0.0
Flask==2.0.2
Flask-Cors==3.0.10
flask-expects-json==1.7.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.5.0
ftfy==6.1.1
future==0.18.2
fvcore==0.1.5.post20221221
gdown==4.7.1
gfpgan==1.3.8
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.59.0
gpustat==1.1.1
gradio==3.41.2
gradio_client==0.5.0
greenlet==2.0.2
grpcio==1.60.0
grpcio-tools==1.48.1
gTTS==2.3.1
gunicorn==21.2.0
h11==0.12.0
h2==4.1.0
hnswlib==0.7.0
hpack==4.0.0
httpcore==0.15.0
httplib2==0.22.0
httptools==0.5.0
httpx==0.24.1
huggingface-hub==0.14.1
humanfriendly==10.0
hypercorn==0.14.4
hyperframe==6.0.1
hyperlink==21.0.0
identify==2.5.22
idna==3.3
imageio==2.33.1
imageio-ffmpeg==0.4.7
importlib-metadata==6.8.0
importlib-resources==6.1.1
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
iopath==0.1.9
isodate==0.6.1
isort==5.12.0
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==1.0.1
joblib==1.2.0
jsonmerge==1.8.0
jsonschema==4.17.3
jsonschema-specifications==2023.6.1
kiwisolver==1.4.4
kornia==0.6.7
langchain==0.0.279
langchainplus-sdk==0.0.20
langsmith==0.0.33
lark==1.1.2
lazy_loader==0.3
librosa==0.8.0
lightning-utilities==0.10.0
linkify-it-py==2.0.2
lion-pytorch==0.1.2
llvmlite==0.41.1
lmdb==1.4.1
local-attention==1.8.6
loguru==0.7.0
looseversion==1.3.0
lpips==0.1.4
lxml==4.9.1
lz4==4.3.2
macholib==1.16.3
mapbox-earcut==1.0.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.0.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.2
mccabe==0.7.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
mediapipe==0.10.9
mmh3==3.1.0
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
msgpack==1.0.5
multidict==6.0.4
musiclm-pytorch==0.2.2
mypy-extensions==1.0.0
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
nltk==3.8.1
nodeenv==1.7.0
npm==0.1.1
numba==0.58.1
numexpr==2.8.4
numpy==1.23.5
nvidia-ml-py==12.535.108
oauth2client==4.1.3
oauthlib==3.2.2
olefile==0.46
omegaconf==2.2.3
onnxruntime==1.16.3
open-clip-torch==2.20.0
openai==0.19.0
openapi-schema-pydantic==1.2.4
opencensus==0.11.2
opencensus-context==0.1.3
opencv-contrib-python==4.8.0.74
opencv-python==4.8.1.78
openpyxl==3.0.9
optional-django==0.1.0
orjson==3.8.10
outcome==1.2.0
overrides==7.3.1
packaging==23.1
pandas==1.3.4
pandas-stubs==1.2.0.35
paramiko==3.3.1
parsel==1.6.0
passlib==1.7.4
pathlib==1.0.1
pathspec==0.11.1
pdf2image==1.16.3
pdfminer.six==20221105
pefile==2021.9.3
pendulum==2.1.2
piexif==1.1.3
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==4.1.0
playsound==1.2.2
pluggy==1.0.0
pooch==1.8.0
portalocker==2.8.2
posthog==3.0.1
pre-commit==3.2.2
prettytable==3.8.0
priority==2.0.0
prometheus-client==0.17.1
Protego==0.2.1
protobuf==3.20.0
prov==2.0.0
psutil==5.9.5
psycopg2-binary==2.9.9
PuLP==2.7.0
pulsar-client==3.2.0
py-spy==0.3.14
pyarrow==13.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycollada==0.7.2
pycparser==2.21
pycryptodome==3.12.0
pydantic==1.10.7
pydantic_core==2.1.2
PyDispatcher==2.0.6
pydot==1.4.2
pydub==0.25.1
pyflakes==3.0.1
pygame==2.1.2
pygit2==1.12.2
Pygments==2.15.1
pyinstaller==5.0
pyinstaller-hooks-contrib==2022.4
pymilvus==2.2.6
pymongo==4.1.0
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pypandoc==1.11
pyparsing==3.0.9
pypdf==3.8.1
PyPDF2==3.0.1
pyquery==1.4.3
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.3.1
pytest-mock==3.10.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.19.2
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytorch-lightning==1.9.4
pytz==2021.3
pytzdata==2020.1
PyWavelets==1.5.0
pywin32-ctypes==0.2.0
pyxnat==1.5
PyYAML==6.0
quart==0.18.4
queuelib==1.6.2
random-user-agent==1.0.1
ray==2.4.0
rdflib==6.3.2
readability-lxml==0.8.1
realesrgan==0.3.0
redis==4.5.4
referencing==0.29.1
regex==2023.5.5
reportlab==4.0.8
repoze.lru==0.7
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.1
resampy==0.4.2
resize-right==0.0.2
rfc3986==1.5.0
rich==13.0.1
rpds-py==0.8.10
rsa==4.7.2
Rtree==1.1.0
runhouse==0.0.11
s3transfer==0.6.2
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.9.3
Scrapy==2.7.0
selenium==4.6.0
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.30.0
service-identity==21.1.0
sgmllib3k==1.0.0
shapely==2.0.2
simplejson==3.19.1
six==1.16.0
skypilot==0.3.3
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.3.2.post1
sourcery==1.2.0
SQLAlchemy==2.0.12
sshfs==2023.7.0
sshtunnel==0.4.0
stability-sdk==0.2.6
starlette==0.26.1
svg.path==6.3
svglib==1.5.1
sympy==1.11.1
tabulate==0.9.0
tb-nightly==2.16.0a20231215
tenacity==8.2.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
tf-keras-nightly==2.16.0.dev2023121510
threadpoolctl==3.1.0
tifffile==2023.12.9
tiktoken==0.4.0
timm==0.9.2
tinycss2==1.2.1
tldextract==3.4.0
tokenizers==0.13.3
tomesd==0.1.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchdiffeq==0.2.3
torchmetrics==1.2.1
torchsde==0.2.6
torchvision==0.15.2
tqdm==4.64.1
traits==6.3.2
trampoline==0.1.2
transformers==4.30.2
trimesh==3.21.7
trio==0.22.0
trio-websocket==0.9.2
tweepy==4.13.0
Twisted==22.8.0
typer==0.7.0
types-pytz==2022.6.0.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.2
ujson==5.7.0
unicodecsv==0.14.1
unstructured==0.8.0
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.22.0
uvloop==0.17.0
vector-quantize-pytorch==1.5.3
virtualenv==20.14.0
w3lib==2.0.1
watchfiles==0.19.0
wcwidth==0.2.6
webdriver-manager==3.8.6
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==2.0.2
wrapt==1.14.1
wsproto==1.2.0
x-clip==0.12.1
xlrd==2.0.1
XlsxWriter==3.1.2
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarl==1.8.2
zipp==3.16.0
zope.interface==5.5.0
zstandard==0.21.0
